new utils class, moved add_editor_bindings to config module
